lets talk about windows
