+++ agent/cmd/nex-agent/main.go
@@ -1,5 +1,3 @@
-//go:build linux
-
 package main
 
 import (
