+++ agent/providers/lib/v8.go
@@ -1,3 +1,5 @@
+//go:build linux && amd64
+
 package lib
 
 import (
