+++ agent/providers/lib/v8_nop.go
@@ -1,0 +1,23 @@
+//go:build !(linux && amd64)
+
+package lib
+
+import (
+	"errors"
+
+	agentapi "github.com/synadia-io/nex/internal/agent-api"
+)
+
+type V8 struct{}
+
+func (V8) Deploy() error { return nil }
+
+func (V8) Execute(subject string, payload []byte) ([]byte, error) { return []byte{}, nil }
+
+func (V8) Undeploy() error { return nil }
+
+func (V8) Validate() error { return nil }
+
+func InitNexExecutionProviderV8(params *agentapi.ExecutionProviderParams) (*V8, error) {
+	return nil, errors.New("V8 is not supported on this platform")
+}
